Break up index.js into multiple component files
A component per file.
Formatted using prettier.
