+++ src/Square.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Square = props => (
+  <button className="square" onClick={props.onClick}>
+    {props.value}
+  </button>
+);
+
+export default Square;
